auth

+++ {{cookiecutter.project_name}}/src/app/api/users/[id]/route.ts
@@ -1,6 +1,8 @@
 import { usersApi } from '@/server/prisma/users'
 import { NextRequest, NextResponse } from 'next/server'
 
+import { privateRoutesMiddleware } from '../../middleware'
+
 interface Params {
   params: {
     id: string
@@ -15,6 +17,9 @@ export async function GET(req: NextRequest, { params: { id } }: Params) {
 }
 
 export async function DELETE(req: NextRequest, { params: { id } }: Params) {
+  const { error: invalidToken } = await privateRoutesMiddleware(req)
+  if (invalidToken) return NextResponse.json({ error: invalidToken.message }, { status: 400 })
+
   const { error } = await usersApi.delete({ id })
 
   if (error) return NextResponse.json({ error: error.message }, { status: 400 })
